+++ .gitignore
@@ -33,3 +33,4 @@ assets/.DS_Store
 data/
 logs/
 scratch/
+pkg/common/config/secrets.go
